fix: 기존 지도 관련 파일 변경
